+ вынес h1, h2 в global styles, правки внешнего вида статистики
