+++ packages/client/src/components/templates/statistics/statistics.styles.tsx
@@ -1,16 +1,6 @@
 import styled from 'styled-components'
 
 export const Section = styled.div`
-  width: 75rem;
-  margin: 0 auto;
-  padding-block: 1.25rem;
+  padding-block: 2.5rem;
   color: var(--color-text);
-`
-
-export const H1 = styled.h1`
-  font-family: var(--font-title);
-  font-size: 1.875rem;
-  line-height: 2.75rem;
-  font-weight: 400;
-  margin-bottom: 1.875rem;
 `
